Remove duplicated cache policy - Now uses `CachePolicy` for both feed and stories cache implementation
