+++ HackrNews/Hackr News Feed Cache/CachePolicy.swift
@@ -6,9 +6,8 @@ import Foundation
 
 struct CachePolicy {
     private static let calendar = Calendar(identifier: .gregorian)
-    private static var maxCacheAgeInDays: Int { 1 }
 
-    static func validate(_ timestamp: Date, against date: Date) -> Bool {
+    static func validate(_ timestamp: Date, against date: Date, maxCacheAgeInDays: Int) -> Bool {
         guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else { return false }
         return date < maxCacheAge
     }
