+++ HackrNews/Hackr News Feed Cache/Story/HackrNewsStoryStore.swift
@@ -4,6 +4,16 @@
 
 import Foundation
 
+public struct CachedStory {
+    public let story: LocalStory
+    public let timestamp: Date
+
+    public init(story: LocalStory, timestamp: Date) {
+        self.story = story
+        self.timestamp = timestamp
+    }
+}
+
 public protocol HackrNewsStoryStore {
     typealias DeletionResult = Swift.Result<Void, Error>
     typealias DeletionCompletion = (DeletionResult) -> Void
@@ -17,51 +27,4 @@ public protocol HackrNewsStoryStore {
     func delete(_ story: LocalStory, completion: @escaping DeletionCompletion)
     func insert(story: LocalStory, with timestamp: Date, completion: @escaping InsertionCompletion)
     func retrieve(completion: @escaping RetrievalCompletion)
-}
-
-public struct CachedStory {
-    public let story: LocalStory
-    public let timestamp: Date
-
-    public init(story: LocalStory, timestamp: Date) {
-        self.story = story
-        self.timestamp = timestamp
-    }
-}
-
-public struct LocalStory: Equatable {
-    public let id: Int
-    public let title: String?
-    public let text: String?
-    public let author: String
-    public let score: Int?
-    public let createdAt: Date
-    public let totalComments: Int?
-    public let comments: [Int]?
-    public let type: String
-    public let url: URL?
-
-    public init(
-        id: Int,
-        title: String?,
-        text: String?,
-        author: String,
-        score: Int?,
-        createdAt: Date,
-        totalComments: Int?,
-        comments: [Int]?,
-        type: String,
-        url: URL?
-    ) {
-        self.id = id
-        self.title = title
-        self.text = text
-        self.author = author
-        self.score = score
-        self.createdAt = createdAt
-        self.totalComments = totalComments
-        self.comments = comments
-        self.type = type
-        self.url = url
-    }
 }
